Minimum and maximum byte requests
Before, a callback could use the “bytes_requested” field to limit the
number of bytes that would be passed in to the process_bytes function.
Now the callback can request a minimum number of bytes, as well.  If
there aren't that many bytes available, the push parser will buffer
the bytes until the minimum is met.