Handle circular callback references when freeing
In more complex parsing callbacks, there can be circular references
amongst a set of callback objects.  If we're not careful about how we
free those callbacks, we can get into an endless loop, since each
callback is responsible for freeing any other callbacks it's connected
to.  Instead of requiring each callback implementation to take care of
breaking circular references, we handle it in push_callback_free().
There's a new “freeing” flag in each push_callback_t instance.  When
we start freeing a callback (but before we call the custom free
function), we set this flag to true.  That way, if the custom free
function tries to free the callback a second time, we can break the
cycle.