Allow parse errors at EOF
We can now detect when a parse error occurs at the end of the data
stream.  Do support this, the code that provides data to the parser
must now call push_parser_eof() once there is no more data in the
stream.  At this point, the push parser will determine whether it's
acceptable for the EOF to occur at this point in the data stream.

There are two possibilities: The first is that we've got some data
leftover from previous calls to push_parser_submit_data().  This data
must be smaller that the current callback's minimum size — if it
weren't, the loop inside of push_parser_submit_data() would've sent
that data into the callback's process_bytes function.  In this case,
we automatically assume that there's a parse error — if the callback
could handle less data successfully, it would've declared that with a
smaller minimum size.

The other possibility is that there's no data left over.  Here, we
have to delegate to the callback to decide if there's a parse error,
since it's application dependent whether an EOF is allowed here.  To
support this, the callback must define a new “eof” function.  The most
common cases will be that EOF is always allowed or never allowed, so
we also define two default implementations for these cases.