Skip callback
This patch adds a callback that skips over a specific number of bytes.
It's similar to the trash callback; however, the trash callback skips
over all remaining bytes, and never passes off to its next callback.
The skip callback, on the other hand, will only skip over the
specified number of bytes.  EOF will not be allowed in the middle of
this sequence — once we've started skipping some of the bytes, we must
skip all of them.