Field callback no longer parses field tag
The field callback is no longer responsible for parsing the field's
tag.  The message callback is going to take care of this, since it
needs to use the field number to dispatch to the right field callback.
The field callback still needs to verify that the wire type is
correct, so it requires someone to fill in the “actual_tag” field
before passing off to the field callback.  The verification routine
will verify the value of this tag.