Tail-calling a callback
This patch adds a new function for “tail-calling” a callback.  This
happens with a bind's second subcallback — once the first callback has
finished, and we've passed its return value into the second callback,
the overall bind behaves exactly as the second callback.
Specifically, when it succeeds, the result of the second callback is
used as the result of the overall bind.  This will happen will several
other combinator callbacks, as well, so we've factored this behavior
out into a separate call.