Fold combinator
This patch adds a “fold” combinator.  For those familiar with Parsec,
fold is similar to many.  For those who aren't,a fold is a loop, which
wraps another callback.  As long as the wrapped callback matches the
input zero or more times, the fold succeeds.  That means that when the
wrapped callback fails at some point, the overall fold succeeds, and
rolls the input back to just before the failed iteration.

This adds a bit of subtlety, since there are certain cases where the
fold can't roll back, because of chunked data.  Specifically, when
can't roll back if the wrapped callback ever returns PUSH_INCOMPLETE —
indicating that the data is okay so far, but that we haven't seen
enough to finish the parse — and then returns PUSH_PARSE_ERROR after
that — indicating that we couldn't decide it was a parse error until
after that chunk boundary.  This is a problem because libpush parsers,
like Parsec parsers, are restricted to LL(1), which means they can't
require any lookahead.

The combinator is called “fold” because it behaves similarly to the
fold function in Haskell.  The input to the fold is used as the input
to the first iteration of the wrapped callback.  Then, the result of
each iteration is passed as input to the next.  The result of the last
successful iteration is used as the result of the overall fold.