Renaming “bind” to “compose”
It turns out this new parser API is equivalent to a Haskell arrow [1],
rather than a monad.  This means that the operator we've been using to
combine parser callbacks isn't the monadic bind operator, it's the
arrow composition operator.  This patch changes the name of the parser
combinator as a result.

[1] John Hughes, “Generalising Monads to Arrows”, in Science of
    Computer Programming 37, pp. 67-111, May 2000.