Removing callback-specific header files
Before, we had a separate header file for each built-in callback
provided by the library.  This was useful when the header files
contained the push_callback_t subclass types, but now that those are
hidden, each header file only defines a single function.  This patch
removes the callback-specific header files, and replaces them with
two: push/primitives.h and push/combinators.h.  The first defines the
callbacks that work on their own; the second defines the callbacks
that wrap other callbacks.  Code that uses the library can still
include push.h to get everything.