Pair callbacks
This patch adds four callbacks for dealing with input and output
values that are pairs: “first”, “second”, “par”, and “both”.  They are
modeled after the first, second, (***), and (&&&) Haskell arrow
operators.  The first and second callbacks apply a wrapped callback to
the one element of a pair.  The par callback does the work of first
and second at the same time, applying separate callbacks to the two
elements of a pair.  The both callback applies two callbacks to a
single input, collecting their results into a pair.