Rewriting test-sum to use pair callbacks
This patch rewrites the test-sum test case to use the new pair
callbacks.  It creates a callback with the following structure:

  +----------------------------------------------+
  |                                              |
  |  NULL  +-----+  uint32_t  +-----+    NULL    |
  |=======>| Int |===========>|     |===========>|
  |        +-----+            | Sum |            |
  |                           |     |            |
  |==========================>|     |===========>|
  |          uint32_t         +-----+  uint32_t  |
  |                                              |
  +----------------------------------------------+

This callback is then folded in the test case to read in a sequence of
integers, summing them together.  The overall callback takes in a
pair, but ignores the first element; the first element is used
internally to store the parsed integer.  The second element of the
pair contains the sum as it accumulates.