Extracting integer callback into separate library
Several of the test cases use the integer callback, so this patch
extracts it out into a library that they all link against.  The
test-int callback had to be modified slightly, since it was testing
the integer callback both with and without its min_bytes wrapper; now,
we just test it with the wrapper.