Extracting sum callback into separate library
This patch extracts the sum callback into a separate library, and
modifies the test-double-sum test case to use it.  The callback has
also been modified; before, it took in a pair and ignored the first
element.  Now, it just takes in a single integer, and constructs the
necessary pair internally.  The new structure of the sum callback is:

  +------------------------------------------------+
  |                                                |
  | ignore +-----+  next int  +-------+            |
  |   /===>| Int |===========>|       |            |
  |==<     +-----+            | Inner |===========>|
  |   \    +------+           |  Sum  |  new sum   |
  |    \==>| Noop |==========>|       |            |
  |  old   +------+  old sum  +-------+            |
  |  sum                                           |
  +------------------------------------------------+