Updating EOF parser to pass input through
The EOF parser now copies its input through to its output if it
successfully parses EOF.  That way, if the EOF parser is composed with
something else

  (foo >>> eof)

the result of the composition will be the result of foo.  The test-eof
test case is updated to reflect this, and now uses the common integer
callback instead of a custom one.