Updating protobuf message callback to use arrow API
The Protocol Buffer message callback is now updated to use the new
arrow-based API.  You now build a “field map”, using basically the
same functions that you used before.  You then create a message
callback from this field map.  Externally, the interface is much the
same; internally, it uses the new combinators to greatly simplify the
design of the callback.