Adding max-bytes combinator
This patch adds a new max-bytes combinator.  This combinator passes
the data it receives into a wrapped callback, but ensures that the
number of bytes passed in doesn't exceed some threshold.  It's okay if
the wrapped callback succeeds or fails before reaching this threshold.
Otherwise, once we reach the threshold, we send in an EOF message to
the wrapped callback, which allows it to determine whether the
threshold is a valid stopping point.