Max-bytes handles incomplete parsing correctly
This patch fixes the behavior of the max-byte combinator when its
wrapped callback yields a PUSH_INCOMPLETE.  First, we were adding the
PUSH_INCOMPLETE code to the data pointer, just like we do with success
codes.  Second, when we were passing EOF in to the wrapped callback,
we returned its result as-is.  This would cause problems if we weren't
at the *real* EOF, since it would cause us to skip over any data
remaining in the buffer.  The test_max_02 test case picks up on both
of these errors.