Adding “dynamic” max-bytes combinator
The existing max-bytes combinator limits the number of bytes sent into
a wrapped callback, but you have to specify a constant number of bytes
when you create the combinator.  This patch adds a “dynamic” version
that takes in the threshold value as an input, which lets you use a
different value each time.  The input to the combinator must be a
pair; the first element is the threshold, and the second element is
the input to the wrapped callback.