Max-bytes callback works on chunked data
Before, the max-bytes callback would fail if its data was sent in
using multiple chunks.  The wrapped callback would return
PUSH_INCOMPLETE at the end of the first chunk, and the max-bytes
wrapper would then treat that as reaching the end of the threshold,
even if there were still bytes left that it could send in.  This patch
adds a check, and only falls through to the “end of threshold” case if
we've actually reached the end of the threshold.