Skipping length-prefixed fields
We want our Protocol Buffer parser to skip over any fields that are
specified in the message callback's field map.  This patch adds
support for skipping over unknown length-prefixed fields.  There is a
separate primitive callback for actually doing the skipping, and the
message callback's dispatcher has been updated to have one of these
lying around that it can dispatch to when it finds a field number that
it doesn't already have a callback for.