Adding callback for reading a string into an HWM buffer
This patch adds a callback that reads a string into an HWM buffer.
The callback doesn't know how to determine the length of the string;
this must be passed in as an input to the callback.  The data will be
parsed into the HWM buffer, and a NUL terminator will be added to the
end of the string.