HWM string callback no longer consumes extra input
The HWM string parser would incorrectly return PUSH_SUCCESS, rather
than the actual number of bytes left over after reading the string.
This would cause the callback to consume everything up through the end
of the current chunk.  This patch fixes this buf, and adds a test case
that includes a string early on in a chunk, which catches the bug.