Making callback pointers const
This patch makes the two function pointers in the push_callback_t type
be const.  (It's the pointers that are const, not the functions
themselves.)  We do the same to the wrapped callback pointers in our
combinator types.  This doesn't have much of a performance impact yet,
but once we start playing with constructing curried callbacks at
runtime, this might allow for some interesting optimizations.