Use separate callbacks in compose combinator
This patch modifies the compose combinator to have two separate
process_bytes functions, depending on which of the wrapped callbacks
is active.  This gives us a slight performance gain over the
single-function implementation, since we get to skip over a test each
time process_bytes is called; but it conflicts with the patch that
makes the callback function pointers const.