Starting continuation-passing implementation
This patch is the start of yet another rewrite of the core API.  The
parser callbacks will now use continuations to link themselves
together.  There are currently four types of continuation:

  • success
  • incomplete
  • error
  • continue

The first three handle the different results that a parser callback
can generate when parsing a chunk of data.  The fourth takes the place
of the old process_bytes function, and is used to resume a callback
that has returned “incomplete”.

Before, callbacks would return a code indicating one of the three
return cases; now, the callback calls one of the three continuations,
instead.  The parser itself defines default implementations of these
callbacks that set the final result of the overall parse.  Combinators
will be implemented by changing the continuations of their wrapped
callbacks, to thread the results through each other as needed.

With this patch, we have the basic API and maintenance functions
defined.  The EOF and noop primitive callbacks have been shifted over
to the new API, as has the integer callback in the test suite.  Many
source files and test cases have been commented out of the build
scripts; they'll be reactivated once they've been migrated over to the
new API.