Min-bytes uses continuation API
This patch updates the minimum-bytes combinator to use the new
continuation API.  It is more complex than the compose implementation,
since has extra behavior both before and after the wrapped callback
executes.  Beforehand, it must ensure that it's received enough data
to meet the minimum; if a chunk comes in that's too small, it must be
buffered until more data comes in.  If we do buffer data, then once
we've met the minimum, we only send in exactly enough data to meet the
minimum.  (This ensures that we don't have to worry about an
expandable buffer.)  However, once we do send in this exact minimum
amount of data, there might still be data left in the current chunk;
this means that once we do pass off to the wrapped callback, we might
have to catch its success and incomplete continuations and modify them
before passing them on to our overall success and incomplete
continuations.