Continuation debug messages
This patch adds new continuation debug messages, which show which
particular function is being called when a continuation is fired.
(Since we were already using a macro to fill in the fields of a
continuation object, it's a trivial modification to include the name
of the function in the continuation object when the debug flag is
turned on.)  This will help us debug compound parsers, to ensure that
the continuations are being threaded together in the expected way.