Consistent result continuation pointers
This patch modifies the existing callbacks so that the pointers to the
success, incomplete, and error callbacks are in the push_callback_t
object.  Before, each callback was responsible for maintaining these
pointers in their user data structs.  The idea was that certain
callbacks wouldn't need all three continuation pointers, since they
might not call them directly.

Unfortunately, this seems to break down with deeply nested compound
callbacks.  For instance, a compose combinator updates its first
wrapped callback so that it succeeds by calling the second callback's
activation function.  If the first callback is itself a combinator,
which doesn't call success on its own, then it probably wouldn't have
a success pointer in its user data struct.  Then, when the compose
tries to update this pointer, to link it with the second callback's
activation function, nothing would happen, and the compose wouldn't
work.

With this patch, all callbacks have pointers to all three
continuations.  Moreover, any combinators that have wrapped callbacks
wait until during the parse to set their continuation pointers.  This
ensures that if the combinator's continuation pointers get changed
while the overall parser is built, those changes are propagated into
the wrapped callbacks correctly.