Fixed callback
This patch adds a new “fixed” callback, which reads in a fixed number
of bytes from the stream.  It uses the minimum-bytes callback to
ensure that this number of bytes is available before trying to parse
the value.  We don't copy the data into a separate buffer; instead,
the result pointer of the parser can point directly into the data
chunk.  This means that consumers of the pointer need to copy the data
themselves, if they need the data to remain across multiple data
chunks.