Max-bytes loops forever when containing a fold
This one takes a bit of effort to encounter.  If you have a fold,
inside a max-bytes, inside another fold, the max-bytes could get into
an endless loop.  The problem comes up when the callback inside the
max-bytes reaches the maximum at the end of a data chunk, but returns
an incomplete at this point.  (A fold is an easy way to do this, if
the folded callback is always going to succeed, since you can't get a
success from the fold until you send it an EOF.)  In this situation,
the max-bytes doesn't detect that the incomplete is at the maximum —
where it should send an EOF — and instead, repeatedly sends in the
last chunk of data, forever.

This patch fixes this, by installing the correct incomplete handler
when the data chunk gives us exactly the maximum.  It also includes a
test case that shows the endless loop behavior without this
modification.