Adding talloc CCAN library
This patch adds the CCAN talloc library [1] to the source tree.  We're
going to use this for memory allocation when creating parsers and
callbacks; this should greatly simplify the process of creating and
freeing these objects.

The only wrinkle is that CCAN libraries are licensed under the LGPL,
whereas libpush has a BSD license.  To make sure we adhere to the
terms of the LGPL, the CCAN code is segregated into a separate
directory, which can easily be replaced with later versions.  The
scons scripts build the ccan code into a static library, which will
then be linked into the libpush code when the final libpush.so library
is created.

[1] http://ccan.ozlabs.org/info/talloc.html