Use talloc to allocate all memory
This patch updates all of the libpush code to use talloc to allocate
memory.  Freeing a push_parser_t should now free all objects
associated with it.  The callback constructors will now return NULL if
any of their wrapped callbacks are NULL.  This makes it easier to
construct new compound callbacks; you can just create the child
callbacks as needed, and check if the final result is NULL.