Use talloc object names instead of callback names
This patch removes the “name” field from the push_callback_t object.
This field was used in debug messages to distinguish between different
instances of the same callback class.  The talloc library allows
objects to have names, as well, and we don't have to jump through as
many hoops when using talloc's naming functions.