Pure function callbacks
This patch adds a new primitive callback that wraps a normal C
function.  The function is given the callback's input pointer, and
should return a bool indicating whether it could calculate an output
value.  If it can, it returns the output pointer, which is immediately
used as the success result for the callback.  Several simpler built-in
callbacks have also been re-written to use the pure callback.