Fixed varint32 slow path for large numbers
The varint32 slow path wouldn't correctly parse any value that
overflowed the 32 bits of the destination field.  Each byte is read
in, and is then shift by the appropriate amount before ORing it into
the current accumulated value.  It looks like the “shift-by” operand
of a left shift is truncated to the number of bits that value can
contain.  So, for instance, a shift-left of 35 (0b100011) is truncated
to a shift left of 3 (0b00011), giving us the incorrect answer.

Our solution is to parse the value into a 64-bit integer in the slow
path, and then to copy this into the 32-bit result field once the
number is fully read.  Since this only happens in the slow path, this
shouldn't cause too much of a performance hit.