Serialization callbacks
It turns out that we don't need to add anything substantial to the
library to support pull serializers — the only real difference between
a push parser and a pull serializer is whether the callback is reading
from or writing to the data chunk.  This patch updates some of the
documentation comments to refer to the possibility of a pull
serializer, and removes the const qualifiers from all of the data
chunk variables and parameters.