Varint32 serialization
This patch adds a callback that serializes a varint32 integer.  Like
the varint32 parser, it has a fast path for when we're activated with
a large enough buffer; in this case, we encode directly into the
output buffer and return.  Otherwise, we encode into a temporary
buffer (which is statically defined to have enough space for any
varint32), and copy from there into the output buffers as they become
available.