Varint serializers take in integer as input
The varint serializers now take in the integer to write as their
input.  Before, they were initialized with an offset, on the
assumption that their input would be some kind of struct that
cnotained an integer; the offset would tell the callback how far into
the struct it should look to find the integer.  I've decided to
separate the concerns more, so now the serializer callbacks just take
in the integer directly.  We'll use some other mechanism to make it
easy to extract and serialize several fields from the same struct.