Arbitrary-length tuples
This patch changes the push_pair_t type to be push_tuple_t, allowing
an arbitrary number of elements per tuple.  The first and second
callbacks have been replaced with a more general nth callback, which
applies a callback to the nth item of a tuple.  (We still have first
and second available as shortcuts for the appropriate nth callbacks.)
The dup callback has also been modified to output a tuple of any
length, rather than only a pair.  The par and both callbacks still
need to be modified; currently, they only operate on 2-tuples.