Varargs constructor for “all” and “par”
The “all” and “par” callbacks now have three separate constructor
functions; they vary in how the list of wrapped callbacks is passed
in.  The “_new” functions take in the list as a variadic argument.
The “_vnew” functions take in a stgargs.h va_list instance.  The
“_anew” functions work as before, taking in a standard C array.  In
all three cases, the elements are expected to be push_callback_t
pointers.